requirements.txt
Flask>=3.1.1
PyYAML>=6.0.2
requests>=2.32.3
gunicorn>=21.2.0